Move response helpers to own file, simplify response objects and update tests
